:memo: Migrate WebSocket client to new structure
